requirements.txt
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4
tzdata==2023.3

djangorestframework~=3.14.0